Checkpoint for IETF 105
Move things around to clean up lots of warnings and hints.
Add some CoRAL work
Remove Json and CBOR link formats.
