+++ ResourceDirectory/Directory.cs
