+++ ResourceDirectory/EndpointLookup.cs
@@ -1,37 +1,38 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Com.AugustCellars.CoAP;
+﻿using System.Text;
+using Com.AugustCellars.CoAP.Coral;
 using Com.AugustCellars.CoAP.Server.Resources;
 using PeterO.Cbor;
-using Com.AugustCellars.CoAP.ResourceDirectory;
 
 namespace Com.AugustCellars.CoAP.ResourceDirectory
 {
-    public class ResourceLookup : Resource
+    public class EndpointLookup : Resource
     {
         private readonly EndpointRegister _root;
 
-        public ResourceLookup(string name, EndpointRegister root) : base(name)
+        public EndpointLookup(string name, EndpointRegister root) : base(name)
         {
             _root = root;
-            Attributes.AddResourceType("core.rd-lookup-res");
+            Attributes.AddResourceType("core.rd-lookup-ep");
             Attributes.AddContentType(MediaType.ApplicationLinkFormat);
+            Attributes.AddContentType(MediaType.ApplicationCoralReef);
+#if false // Dead code
             Attributes.AddContentType(MediaType.ApplicationCbor);
             Attributes.AddContentType(MediaType.ApplicationJson);
-            root.ResourceLookupResource = this;
-        }
+#endif
 
+            root.EndpointLookupResource = this;
+        }
 
         protected override void DoGet(CoapExchange exchange)
         {
             Request req = exchange.Request;
             CBORObject items = null;
             StringBuilder sb = null;
+#if false // dead
             Dictionary<string, CBORObject> dict = null;
-            int respContentType = MediaType.ApplicationLinkFormat;
+#endif
+            int retContentType = MediaType.ApplicationLinkFormat;
+            CoralBody coral = null;
 
             try {
                 Filter filter = new Filter(req.UriQueries);
@@ -43,7 +44,8 @@ namespace Com.AugustCellars.CoAP.ResourceDirectory
                     lastItem = firstItem + filter.Page - 1;
                 }
                 else if (filter.Page != 0) {
-                    throw new Exception();
+                    exchange.Respond(StatusCode.BadRequest);
+                    return;
                 }
 
                 Response resp = Response.CreateResponse(req, StatusCode.Content);
@@ -55,6 +57,7 @@ namespace Com.AugustCellars.CoAP.ResourceDirectory
                             sb = new StringBuilder();
                             break;
 
+#if false // Dead work
                         case MediaType.ApplicationLinkFormatCbor:
                             items = CBORObject.NewArray();
                             dict = LinkFormat.CborAttributeKeys;
@@ -62,6 +65,11 @@ namespace Com.AugustCellars.CoAP.ResourceDirectory
 
                         case MediaType.ApplicationLinkFormatJson:
                             items = CBORObject.NewArray();
+                            break;
+#endif
+
+                        case MediaType.ApplicationCoralReef:
+                            coral = new CoralBody();
                             break;
 
                         default:
@@ -71,19 +79,20 @@ namespace Com.AugustCellars.CoAP.ResourceDirectory
                         }
 
                         //  We found a value
-                        if (sb != null || items != null) {
-                            respContentType = acceptOption.IntValue;
+                        if (sb != null || items != null || coral != null) {
+                            retContentType = acceptOption.IntValue;
                             break;
                         }
                     }
 
-                    if (sb == null && items == null) {
+                    if (sb == null && items == null && coral == null) {
                         exchange.Respond(StatusCode.NotAcceptable);
                         return;
                     }
                 }
                 else {
                     sb = new StringBuilder();
+                    retContentType = MediaType.ApplicationLinkFormat;
                 }
 
                 int itemCount = -1;
@@ -91,41 +100,43 @@ namespace Com.AugustCellars.CoAP.ResourceDirectory
                 foreach (EndpointNode ep in _root.ChildEndpointNodes) {
                     if (ep.IsDeleted) continue;
 
-                    Uri baseUri = ep.BaseUrl;
-
-                    foreach (IResource link in ep.Links) {
-                        filter.ClearState();
-                        if (!filter.Apply(link.Attributes)) {
-                            if (!ep.ApplyFilter(filter, true, false)) {
-                                filter.Href(link.Uri);
-                            }
+                    filter.ClearState();
+                    ep.ApplyFilter(filter, true, true);
+                    if (filter.Passes) {
+                        itemCount += 1;
+                        if (itemCount < firstItem || itemCount > lastItem) continue;
+                        if (coral != null) {
+                            ep.GetLink(coral);
                         }
-                        if (filter.Passes) {
-                            itemCount += 1;
-                            if (itemCount < firstItem || itemCount > lastItem) continue;
-                            if (sb != null) {
-                                LinkFormat.SerializeResource(link, sb, null, baseUri);
-                                sb.Append(',');
-                            }
-                            else {
-                                LinkFormat.SerializeResource(link, items, dict, null, baseUri);
-                            }
+                        else if (sb != null) {
+                            ep.GetLink(sb);
                         }
+                        
+#if false  // Work is dead?
+                        else {
+                            ep.GetLink(items, dict, retContentType);
+                        }
+#endif
                     }
                 }
 
-                if (sb != null) {
+                if (coral != null) {
+                    resp.Payload = coral.EncodeToBytes(LinkFormat.ReefDictionary);
+                }
+                else if (sb != null) {
                     if (sb.Length > 0) sb.Remove(sb.Length - 1, 1);
                     resp.PayloadString = sb.ToString();
                 }
+#if false // Dead?
                 else if (dict == null) {
                     resp.PayloadString = items.ToJSONString();
                 }
                 else {
                     resp.Payload = items.EncodeToBytes();
                 }
+#endif
 
-                resp.ContentFormat = respContentType;
+                resp.ContentFormat = retContentType;
 
                 exchange.Respond(resp);
             }
