+++ ResourceDirectory/Properties/AssemblyInfo.cs
