+++ ResourceDirectory/RD.dev.csproj
