+++ ResourceDirectory/ResourceLookup.cs
@@ -4,22 +4,28 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Com.AugustCellars.CoAP;
+using Com.AugustCellars.CoAP.Coral;
 using Com.AugustCellars.CoAP.Server.Resources;
 using PeterO.Cbor;
-#if false
+using Com.AugustCellars.CoAP.ResourceDirectory;
+using Com.AugustCellars.CoAP.Util;
 
 namespace Com.AugustCellars.CoAP.ResourceDirectory
 {
-    public class GroupLookup : Resource
+    public class ResourceLookup : Resource
     {
-        private readonly GroupManager _groupManager;
-        public GroupLookup(string name, GroupManager groupManager, EndpointRegister rd) : base(name)
-        {
-            Attributes.AddResourceType("core.rd-lookup-gp");
-            Attributes.AddContentType(MediaType.ApplicationLinkFormat);
+        private readonly EndpointRegister _root;
 
-            _groupManager = groupManager;
+        public ResourceLookup(string name, EndpointRegister root) : base(name)
+        {
+            _root = root;
+            Attributes.AddResourceType("core.rd-lookup-res");
+            Attributes.AddContentType(MediaType.ApplicationLinkFormat);
+            Attributes.AddContentType(MediaType.ApplicationCbor);
+            Attributes.AddContentType(MediaType.ApplicationJson);
+            root.ResourceLookupResource = this;
         }
+
 
         protected override void DoGet(CoapExchange exchange)
         {
@@ -28,6 +34,7 @@ namespace Com.AugustCellars.CoAP.ResourceDirectory
             StringBuilder sb = null;
             Dictionary<string, CBORObject> dict = null;
             int respContentType = MediaType.ApplicationLinkFormat;
+            CoralBody coral = null;
 
             try {
                 Filter filter = new Filter(req.UriQueries);
@@ -39,8 +46,7 @@ namespace Com.AugustCellars.CoAP.ResourceDirectory
                     lastItem = firstItem + filter.Page - 1;
                 }
                 else if (filter.Page != 0) {
-                    exchange.Respond(StatusCode.BadRequest);
-                    return;
+                    throw new Exception();
                 }
 
                 Response resp = Response.CreateResponse(req, StatusCode.Content);
@@ -52,6 +58,7 @@ namespace Com.AugustCellars.CoAP.ResourceDirectory
                             sb = new StringBuilder();
                             break;
 
+#if false  // Work is dead?
                         case MediaType.ApplicationLinkFormatCbor:
                             items = CBORObject.NewArray();
                             dict = LinkFormat.CborAttributeKeys;
@@ -59,6 +66,11 @@ namespace Com.AugustCellars.CoAP.ResourceDirectory
 
                         case MediaType.ApplicationLinkFormatJson:
                             items = CBORObject.NewArray();
+                            break;
+#endif
+
+                        case 65088:
+                                coral = new CoralBody();
                             break;
 
                         default:
@@ -68,43 +80,63 @@ namespace Com.AugustCellars.CoAP.ResourceDirectory
                         }
 
                         //  We found a value
-                        if (sb != null || items != null) {
+                        if (sb != null || items != null || coral != null) {
                             respContentType = acceptOption.IntValue;
                             break;
                         }
                     }
 
-                    if (sb == null && items == null) {
+                    if (sb == null && items == null && coral == null) {
                         exchange.Respond(StatusCode.NotAcceptable);
                         return;
                     }
                 }
                 else {
                     sb = new StringBuilder();
-                    resp.ContentFormat = MediaType.ApplicationLinkFormat;
                 }
 
                 int itemCount = -1;
 
-                foreach (GroupLeaf ep in _groupManager.AllGroups) {
-                    filter.ClearState();
-                    ep.ApplyFilter(filter, true);
+                foreach (EndpointNode ep in _root.ChildEndpointNodes) {
+                    if (ep.IsDeleted) continue;
 
-                    if (filter.Passes) {
-                        itemCount += 1;
-                        if (itemCount < firstItem || itemCount > lastItem) continue;
-                        if (sb != null) {
-                            ep.SerializeResource(sb);
-                            sb.Append(',');
+                    Uri baseUri = ep.BaseUrl;
+
+                    if (respContentType == 65088) {
+                        coral.Add(new CoralBase(baseUri));
+                    }
+
+                    foreach (IResource link in ep.Links) {
+                        filter.ClearState();
+                        if (!filter.Apply(link.Attributes)) {
+                            if (!ep.ApplyFilter(filter, true, false)) {
+                                filter.Href(link.Uri);
+                            }
                         }
-                        else {
-                            ep.SerializeResource(items, dict);
+                        if (filter.Passes) {
+                            itemCount += 1;
+                            if (itemCount < firstItem || itemCount > lastItem) continue;
+                            if (sb != null) {
+                                LinkFormat.SerializeResource(link, sb, null, baseUri);
+                                sb.Append(',');
+                            }
+                            else if (respContentType == MediaType.ApplicationCoralReef) {
+
+                                LinkFormat.SerializeResourceInCoral(link, coral, dict, null, null);
+                            }
+#if false  // Work is dead?
+                            else { 
+                                LinkFormat.SerializeResource(link, items, dict, null, baseUri);
+                            }
+#endif
                         }
                     }
                 }
 
-
-                if (sb != null) {
+                if (coral != null) {
+                    resp.Payload = coral.EncodeToBytes(LinkFormat.ReefDictionary);
+                }
+                else if (sb != null) {
                     if (sb.Length > 0) sb.Remove(sb.Length - 1, 1);
                     resp.PayloadString = sb.ToString();
                 }
@@ -125,4 +157,3 @@ namespace Com.AugustCellars.CoAP.ResourceDirectory
         }
     }
 }
-#endif
