+++ ResourceDirectory/WellKnown.cs
