+++ ResourceDirectory/packages.config
