+++ foo/ResourceDirectory.csproj
