+++ foo/ResourceDirectoryResource.cs
